Move component commands into the component directory
