+++ test/component/cypress/support/index.js
@@ -1,3 +1,4 @@
 /* eslint-disable */
 import '@cypress/code-coverage/support'
-require('../../cypress/support/commands')
+require('../../../cypress/support/commands')
+require('./commands')
