+++ test/cypress/support/commands.js
@@ -14,17 +14,6 @@
  * the log message.
  * @returns undefined
  */
-import React from 'react'
-import { mount } from 'cypress/react'
-
-import { createTestProvider } from '../../../test/component/cypress/specs/provider'
-import { TASK_GET_REMINDER_SUMMARY } from '../../../src/client/components/NotificationAlert/state'
-import { INITIAL_STATE } from '../../../src/client/components/NotificationAlert/reducer'
-import { TASK_GET_TYPEAHEAD_OPTIONS } from  '../../../src/client/components/Typeahead/state'
-import {getTypeaheadOptions } from  '../../../src/client/components/Typeahead/tasks'
-
-Cypress.Commands.add('mount', mount)
-
 const addLoggedCommand = ({
   name,
   logName,
@@ -375,18 +364,3 @@ Cypress.Commands.add('isScrolledTo', (element) => {
 
 Cypress.Commands.add('dataTest', (value) => cy.get(`[data-test="${value}"]`))
 
-Cypress.Commands.add(
-  'mountWithProvider',
-  (children, { tasks, initialPath } = {}) => {
-    const Provider = createTestProvider({
-      tasks: {
-        [TASK_GET_REMINDER_SUMMARY]: () => Promise.resolve(INITIAL_STATE),
-        [TASK_GET_TYPEAHEAD_OPTIONS]: getTypeaheadOptions,
-        ...tasks,
-      },
-      initialPath,
-    })
-    
-    return cy.mount(<Provider>{children}</Provider>)
-  }
-)
